Pre-calculate king attacks. Bench: 14632975
